requirements.txt
Flask==2.0.1
Pillow==8.3.1
joblib==1.0.1
pandas==1.3.0
scikit-learn==0.24.2